create AggrMethods.ts for supporting aggregation functionality
